Add new solution class for steady state algorithms
