+++ openpnm/algorithms/_reactive_transport.py
@@ -4,6 +4,7 @@ from scipy.optimize.nonlin import TerminationCondition
 from openpnm.algorithms import GenericTransport
 from openpnm.utils import logging
 from openpnm.utils import TypedList, Docorator, SettingsAttr
+from ._solution import SteadyStateSolution
 docstr = Docorator()
 logger = logging.getLogger(__name__)
 
@@ -236,17 +237,22 @@ class ReactiveTransport(GenericTransport):
         xold = self.x
         dx = self.x - xold
         condition = TerminationCondition(f_rtol=f_rtol, x_rtol=x_rtol)
+        # Store the solution as a SteadyStateSolution object on algorithm
+        self.soln = SteadyStateSolution(self.x)
 
         for i in range(maxiter):
+            self.soln.num_iter = i + 1
             res = self._get_residual()
-            self.is_converged = condition.check(f=res, x=xold, dx=dx)
-            if self.is_converged:
+            self.soln.is_converged = bool(condition.check(f=res, x=xold, dx=dx))
+            if self.soln.is_converged:
                 logger.info(f'Solution converged, residual norm: {norm(res):.4e}')
                 return
             super()._run_special(solver=solver, x0=xold, w=w)
             dx = self.x - xold
             xold = self.x
+            self.soln[:] = self.x
             logger.info(f'Iteration #{i:<4d} | Residual norm: {norm(res):.4e}')
+
         logger.critical(f"{self.name} didn't converge after {maxiter} iterations")
 
     def _update_A_and_b(self):
