+++ openpnm/integrators/_scipy.py
@@ -1,9 +1,8 @@
-import numpy as np
 from scipy.integrate import solve_ivp
-from scipy.interpolate import interp1d
 from openpnm.integrators import Integrator
+from openpnm.algorithms._solution import TransientSolution
 
-__all__ = ['ScipyRK45', 'Solution', 'TransientSolution']
+__all__ = ['ScipyRK45']
 
 
 class ScipyRK45(Integrator):
@@ -54,46 +53,3 @@ class ScipyRK45(Integrator):
         if sol.success:
             return TransientSolution(sol.t, sol.y)
         raise Exception(sol.message)
-
-
-class Solution(np.ndarray):
-    """Brief description of 'Solution'"""
-    ...
-
-
-class TransientSolution(Solution):
-    """Brief description of 'TransientSolution'"""
-
-    def __new__(cls, t, x):
-        obj = np.asarray(x).view(cls)
-        obj.t = np.asarray(t).view(cls)
-        return obj
-
-    def _create_interpolant(self):
-        self._interpolant = interp1d(self.t, self, bounds_error=True)
-
-    def interpolate(self, t):
-        """
-        Interpolates solution at time 't'.
-
-        Parameters
-        ----------
-        t : float
-            Time at which the solution is to be interpolated
-
-        Returns
-        -------
-        ndarray
-            Transient solution interpolated at the given time 't'
-
-        Notes
-        -----
-        't' must reside inside the 'tspan' used during integration.
-
-        """
-        # Cache interpolant to avoid overhead
-        if not hasattr(self, "_interpolant"):
-            self._create_interpolant()
-        return self._interpolant(t)
-
-    __call__ = interpolate
